+++ packages/agents-core/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @inkeep/agents-core
 
+## 0.16.0
+
+### Minor Changes
+
+- 5c3bbec: Request context refactor
+
+### Patch Changes
+
+- 35e6c9e: Updated Artifact Schema
+
+## 0.15.0
+
+### Minor Changes
+
+- ad5528c: Context config route changes
+
 ## 0.14.16
 
 ## 0.14.15