+++ packages/create-agents/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @inkeep/create-agents
 
+## 0.16.0
+
+## 0.15.0
+
 ## 0.14.16
 
 ### Patch Changes